requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==3.0.2
nulltype==2.3.1
oauthlib==3.2.2
packaging==24.2
plaid-python==28.0.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==0.15.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
